pyproject.toml
[tool.poetry]
name = "ml_gcp_api_labs_back_end"
version = "0.1.0"
description = ""
authors = ["Smar-Ben <benrams1997@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
google-cloud-language = "^2.6.1"
google-cloud-texttospeech = "^2.12.3"
google-cloud-translate = "^3.8.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
